限流修改

+++ route-server-zuul/src/main/resources/h2/schema.sql
@@ -4,7 +4,7 @@ CREATE TABLE `audit_log`
     `username` varchar(64) COMMENT '用户名',
     `path`     varchar(64) COMMENT '路径',
     `method`   varchar(64) COMMENT '方法',
-    `status`   tinyint(4) COMMENT '状态',
+    `status`   smallint COMMENT '状态',
     `create_time`   datetime COMMENT '创建时间',
     `update_time`   datetime COMMENT '修改时间',
     PRIMARY KEY (`id`)
